fix(chat): do not close chat window when conversation has ended
